+++ lib/src/steps/predefined_steps/completion_step.dart
@@ -14,11 +14,14 @@ class CompletionStep extends Step {
 
   CompletionStep({
     bool isOptional = false,
-    required StepIdentifier id,
+    required StepIdentifier stepIdentifier,
     String buttonText = 'End Survey',
     required this.title,
     required this.text,
-  }) : super(id: id, isOptional: isOptional, buttonText: buttonText);
+  }) : super(
+            stepIdentifier: stepIdentifier,
+            isOptional: isOptional,
+            buttonText: buttonText);
 
   @override
   Widget createView({required QuestionResult? questionResult}) {