+++ lib/src/steps/step.dart
@@ -6,14 +6,17 @@ import 'package:survey_kit/src/steps/step_not_defined_exception.dart';
 import 'package:survey_kit/survey_kit.dart';
 
 abstract class Step {
-  final StepIdentifier id;
+  final StepIdentifier stepIdentifier;
   @JsonKey(defaultValue: false)
   final bool isOptional;
   @JsonKey(defaultValue: 'Next')
   final String buttonText;
 
-  Step({StepIdentifier? id, this.isOptional = false, this.buttonText = 'Next'})
-      : id = id ?? StepIdentifier();
+  Step(
+      {StepIdentifier? stepIdentifier,
+      this.isOptional = false,
+      this.buttonText = 'Next'})
+      : stepIdentifier = stepIdentifier ?? StepIdentifier();
 
   Widget createView({required QuestionResult? questionResult});